+++ design-system/page-layout/src/controllers/sidebar-resize-context.tsx
@@ -1,4 +1,10 @@
-import { createContext, KeyboardEvent, MouseEvent, useContext } from 'react';
+import {
+  createContext,
+  KeyboardEvent,
+  MouseEvent,
+  useContext,
+  useEffect,
+} from 'react';
 
 const noop = () => {};
 
@@ -8,7 +14,9 @@ export type LeftSidebarState = {
   isLeftSidebarCollapsed: boolean;
   leftSidebarWidth: number;
   lastLeftSidebarWidth: number;
+  flyoutLockCount: number;
 };
+
 export type SidebarResizeContextValue = {
   isLeftSidebarCollapsed: boolean;
   expandLeftSidebar: () => void;
@@ -17,7 +25,11 @@ export type SidebarResizeContextValue = {
     collapseWithoutTransition?: boolean,
   ) => void;
   leftSidebarState: LeftSidebarState;
-  setLeftSidebarState: (leftSidebarState: LeftSidebarState) => void;
+  setLeftSidebarState: (
+    value:
+      | LeftSidebarState
+      | ((prevState: LeftSidebarState) => LeftSidebarState),
+  ) => void;
 };
 
 const leftSidebarState = {
@@ -26,6 +38,7 @@ const leftSidebarState = {
   isLeftSidebarCollapsed: false,
   leftSidebarWidth: 0,
   lastLeftSidebarWidth: 0,
+  flyoutLockCount: 0,
 };
 export const SidebarResizeContext = createContext<SidebarResizeContextValue>({
   isLeftSidebarCollapsed: false,
@@ -36,5 +49,41 @@ export const SidebarResizeContext = createContext<SidebarResizeContextValue>({
 });
 
 export const usePageLayoutResize = () => {
-  return useContext(SidebarResizeContext);
+  const { setLeftSidebarState, ...context } = useContext(SidebarResizeContext);
+  return context;
+};
+
+/**
+ * **WARNING:** This hook is intended as a temporary solution and
+ * is likely to be removed in a future version of page-layout.
+ *
+ * ---
+ *
+ * This hook will prevent the left sidebar from automatically collapsing
+ * when it is in a flyout state.
+ *
+ * The intended use case for this hook is to allow popup menus in the
+ * left sidebar to be usable while it is in a flyout state.
+ *
+ * ## Usage
+ * The intended usage is to use this hook within the popup component
+ * you are rendering. This way the left sidebar will be locked for
+ * as long as the popup is open.
+ */
+
+export const useLeftSidebarFlyoutLock = () => {
+  const { setLeftSidebarState } = useContext(SidebarResizeContext);
+
+  useEffect(() => {
+    setLeftSidebarState(current => ({
+      ...current,
+      flyoutLockCount: current.flyoutLockCount + 1,
+    }));
+    return () => {
+      setLeftSidebarState(current => ({
+        ...current,
+        flyoutLockCount: current.flyoutLockCount - 1,
+      }));
+    };
+  }, [setLeftSidebarState]);
 };